removed redirects from printOrder
